requirements/backend.txt
flask==1.1.2
Flask-SQLAlchemy==2.4.4
psycopg2==2.8.6
Flask-Migrate==2.5.3
pytest==6.1.0
flask-restful==0.3.8
flask-marshmallow==0.14.0
marshmallow-sqlalchemy==0.23.1
pyjwt==1.7.1
